refactor: use `sidekick-new`
