+++ src/components/messenger/list/index.tsx
@@ -29,7 +29,7 @@ import { IconChevronLeft, IconChevronRight } from '@zero-tech/zui/icons';
 import { ErrorDialog } from '../../error-dialog';
 import { ErrorDialogContent } from '../../../store/chat/types';
 import { receiveSearchResults } from '../../../store/users';
-import { CurrentUserDetails } from '../../sidekick/components/current-user-details';
+import { CurrentUserDetails } from '../../sidekick-new/components/current-user-details';
 import { getUserSubHandle } from '../../../lib/user';
 import { VerifyIdDialog } from '../../verify-id-dialog';
 import { RewardsModalContainer } from '../../rewards-modal/container';
@@ -39,7 +39,8 @@ import { Button } from '@zero-tech/zui/components/Button';
 import { IconPlus } from '@zero-tech/zui/icons';
 import { GroupTypeDialog } from './group-details-panel/group-type-dialog';
 import { AdminMessageType } from '../../../store/messages';
-import { Header } from '../../sidekick/components/header';
+import { Header } from '../../sidekick-new/components/header';
+import { Content as SidekickContent } from '../../sidekick-new/components/content';
 
 import { bemClassName } from '../../../lib/bem';
 import './styles.scss';
@@ -348,13 +349,13 @@ export class Container extends React.Component<Properties, State> {
     return (
       <>
         {this.props.stage === SagaStage.None && this.renderUserHeader()}
-        <div {...cn('')}>
+        <SidekickContent>
           {this.renderCreateConversation()}
           {this.state.isVerifyIdDialogOpen && this.renderVerifyIdDialog()}
           {this.props.joinRoomErrorContent && this.renderErrorDialog()}
           {this.props.isRewardsDialogOpen && this.renderRewardsDialog()}
           {this.state.isGroupTypeDialogOpen && this.renderGroupTypeDialog()}
-        </div>
+        </SidekickContent>
         {isExpanded && this.props.stage === SagaStage.None && this.renderFooterButton()}
         {this.renderInviteDialog()}
       </>
